requirements.txt
colorlog==6.8.2
discord.py==2.3.2
requests==2.28.2
humanize==4.9.0